Major usability update.
- Workflow template now expands copy of the official Bevy GH template.
- Binary template refactored to be a little clearer.
- Binary template setup to support `wasm-server-runner`.
- Workflow setup includes WASM support.
- Readme updated to reflect state of project.
