+++ Binary/cargo-generate.toml
@@ -12,11 +12,9 @@ camera = { type = "string", prompt = "Camera type?", choices = [
     "3D",
 ], default = "2D" }
 
-[conditional.'within_cargo_project == false'.placeholders]
-ide = { type = "string", prompt = "Which IDE to use?", default = "vscode", choices = ["none", "vscode"] }
-
 [conditional.'within_cargo_project']
 ignore = [
+    ".cargo",
     ".github",
     ".vscode",
     ".gitignore",
@@ -24,6 +22,9 @@ ignore = [
 
 [conditional.'within_cargo_project && is_init']
 ignore = ["README.md", "Cargo.toml"]
+
+[conditional.'within_cargo_project == false'.placeholders]
+ide = { type = "string", prompt = "Which IDE to use?", default = "vscode", choices = ["none", "vscode"] }
 
 [conditional.'within_cargo_project == false && ide != "vscode"']
 ignore = [".vscode"]
