+++ Library/.vscode/tasks.json
@@ -7,6 +7,7 @@
       "label": "Test (fast compile)",
       "detail": "Runs all tests",
       "type": "shell",
+      "group": "test",
       "command": "cargo t --workspace --features bevy/dynamic",
       "options": {
         "env": {
@@ -14,7 +15,6 @@
         }
       },
       "problemMatcher": ["$rustc"],
-      "group": "build",
       "presentation": {
         "echo": true,
         "reveal": "always",
@@ -28,9 +28,9 @@
       "label": "Build release",
       "detail": "Make a release build",
       "type": "shell",
+      "group": "build",
       "command": "cargo build --release",
       "problemMatcher": ["$rustc"],
-      "group": "build",
       "presentation": {
         "echo": true,
         "reveal": "always",
@@ -44,9 +44,9 @@
       "label": "Build debug",
       "detail": "Make a debug build",
       "type": "shell",
+      "group": "build",
       "command": "cargo build",
       "problemMatcher": ["$rustc"],
-      "group": "build",
       "presentation": {
         "echo": true,
         "reveal": "always",
@@ -55,14 +55,6 @@
         "clear": true,
         "focus": true
       }
-    },
-    {
-      "label": "Open doc",
-      "detail": "Open the documentation",
-      "type": "shell",
-      "command": "cargo doc --open",
-      "problemMatcher": ["$rustc"],
-      "dependsOn": "Build doc"
     },
     {
       "label": "Build doc",
@@ -81,20 +73,12 @@
       }
     },
     {
-      "label": "Check all-features (fast compile)",
-      "detail": "Run a compile check against all features",
+      "label": "Open doc",
+      "detail": "Open the documentation",
       "type": "shell",
-      "command": "cargo check --features bevy/dynamic --all-features --all-targets",
+      "command": "cargo doc --open",
       "problemMatcher": ["$rustc"],
-      "group": "build",
-      "presentation": {
-        "echo": true,
-        "reveal": "always",
-        "panel": "shared",
-        "showReuseMessage": true,
-        "clear": true,
-        "focus": true
-      }
+      "dependsOn": "Build doc"
     },
     {
       "label": "Update",
@@ -131,6 +115,19 @@
       "detail": "Format all source",
       "type": "shell",
       "command": "cargo fmt --all --quiet",
+      "problemMatcher": [],
+      "presentation": {
+        "echo": false,
+        "reveal": "never",
+        "panel": "shared",
+        "showReuseMessage": false
+      }
+    },
+    {
+      "label": "Clean",
+      "detail": "Clean build artifacts",
+      "type": "shell",
+      "command": "cargo clean",
       "problemMatcher": [],
       "presentation": {
         "echo": false,
