+++ Workflow/final-msg.rhai
@@ -1,0 +1,3 @@
+print();
+print("Remember to check the name of the binary in .github/workflows/release.yaml");
+print();
