+++ Workflow/wasm/index.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html lang="en">
+<script type="module">
+  import init from './bevy_game.js'
+  init()
+</script>
+
+<body style="margin: 0px;">
+</body>
+
+</html>
